requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.3
cryptography==45.0.5
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docx==0.2.4
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.59.1
fpdf2==2.8.4
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.30.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
llvmlite==0.43.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
packaging==25.0
pandas==2.3.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.31
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
whisper==1.1.10
zipp==3.23.0
